Updated resolveReferences function
This change integrates code from the web payments activity that handles
the dropping of unused definitions in collections of common terms.  It
addresses a bug where if term A was referenced from the main document,
and term A referenced term B but term B was NOT referenced by the main
document, term B would be removed even though it is needed.